Modularize client spec file in a client directory
