+++ test/client/client.spec.js
@@ -1,0 +1,5 @@
+import { GridActions } from './actions/grid'
+
+describe('Actions', () => {
+  describe('Grid Actions', GridActions)
+})
